server/requirements.txt
flask
SpeechRecognition
python-socketio
eventlet
pydub
